examples/requirements/testing.txt
beautifulsoup4==4.5.1
factory_boy==2.7.0
fake-factory==0.7.2
Pillow==3.4.2
pytest==3.0.2
pytest-django==2.9.1
pytest-cov==2.2.1
tox==2.1.1
selenium==2.53.6
